Start separating CCMPS model from UI code
* Start extracting data model from UI code
* Merge `CCMPS` and `CCMPS_actions`
* No change in functionality

Related to #46.
